split io::read into multiple files
